# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaign-drafts:
    post:
      operationId: createCampaignDraft
      tags: [campaigns]
      summary: Create an unpublished, expiring campaign draft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignDraftInput" }
      responses:
        "201":
          description: Draft created and owner capability cookie issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignDraft" }
        "400": { description: Invalid draft }
        "409": { description: Campaign ID already exists }
        "429": { description: Too many requests }
  /campaign-drafts/{campaignId}/w9/request-url:
    post:
      operationId: requestCampaignDraftW9Upload
      tags: [campaigns, storage]
      summary: Issue a W-9 upload intent for an authorized high-value draft
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/W9UploadRequest" }
      responses:
        "201":
          description: Single-use W-9 upload intent and signed URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/W9UploadIntent" }
        "400": { description: Invalid W-9 metadata or ineligible draft }
        "404": { description: Draft not found or not authorized }
        "429": { description: Too many requests }
  /campaign-drafts/{campaignId}/w9/{intentId}/finalize:
    post:
      operationId: finalizeCampaignDraftW9Upload
      tags: [campaigns, storage]
      summary: Verify and finalize a W-9 upload intent once
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - name: intentId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: W-9 finalized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/W9UploadIntent" }
        "400": { description: Uploaded W-9 is missing or does not match intent metadata }
        "404": { description: Draft or intent not found or not authorized }
        "409": { description: "Intent already finalized, consumed, or expired" }
  /campaign-drafts/{campaignId}/publish:
    post:
      operationId: publishCampaignDraft
      tags: [campaigns]
      summary: Publish an authorized campaign draft
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublishCampaignDraftInput" }
      responses:
        "200":
          description: Published campaign
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "400": { description: W-9 or social context is required }
        "404": { description: Draft not found or not authorized }
        "409": { description: "Draft expired, already published, or W-9 intent reused" }
...
components:
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
    CampaignId:
      in: path
      name: campaignId
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
  schemas:
...
    CampaignDraftInput:
      type: object
      required: [id, itemType, title, ownerName, ownerEmail, pricesCents, ownerAssent, presentation]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        itemType: { type: string, minLength: 1, maxLength: 50 }
        title: { type: string, minLength: 1, maxLength: 200 }
        ownerName: { type: string, minLength: 1, maxLength: 120 }
        ownerEmail: { type: string, format: email, maxLength: 320 }
        pricesCents:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: integer, minimum: 100, maximum: 10000000 }
        ownerAssent: { $ref: "#/components/schemas/OwnerAssent" }
        presentation: { type: object, maxProperties: 32, additionalProperties: true }
    CampaignDraft:
      type: object
      required: [id, status, expiresAt, w9Required]
      properties:
        id: { type: string }
        status: { type: string, enum: [draft] }
        expiresAt: { type: string, format: date-time }
        w9Required: { type: boolean }
    W9UploadRequest:
      type: object
      additionalProperties: false
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1, maximum: 10000000 }
        contentType: { type: string, const: application/pdf }
    W9UploadIntent:
      type: object
      required: [id, purpose, objectPath, expectedMimeType, expectedSizeBytes, expectedFileName, status, expiresAt]
      properties:
        id: { type: string }
        purpose: { type: string, const: campaign_draft_w9 }
        objectPath: { type: string, pattern: "^/objects/uploads/[A-Za-z0-9-]+$" }
        expectedMimeType: { type: string, const: application/pdf }
        expectedSizeBytes: { type: integer }
        expectedFileName: { type: string }
        status: { type: string, enum: [issued, finalized, consumed, revoked] }
        expiresAt: { type: string, format: date-time }
        uploadURL: { type: string, format: uri }
    PublishCampaignDraftInput:
      type: object
      additionalProperties: false
      properties:
        w9IntentId: { type: string, minLength: 1 }
    Campaign:
      type: object
      required: [id, itemType, title, ownerName, pricesCents, presentation, claims, active, lifecycleStatus, relistCount, relistEligible, createdAt]
      properties:
        id:
          type: string
        itemType:
          type: string
        title:
          type: string
        ownerName:
          type: string
        pricesCents:
          type: array
          items:
            type: integer
        presentation:
          type: object
          maxProperties: 32
          additionalProperties: true
        claims:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/PublicClaim"
              - type: "null"
        active:
          type: boolean
        lifecycleStatus:
          type: string
          enum: [live, funding, funded, ordered, branded, shipped, active, complete, expired]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        fundedAt:
          type: ["string", "null"]
          format: date-time
        relistCount:
          type: integer
        relistEligible:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    OwnerAssent:
      type: object
      required: [accepted, termsVersion, contentVersion, checkinVersion]
      properties:
        accepted:
          type: boolean
          const: true
        termsVersion:
          type: string
        contentVersion:
          type: string
        checkinVersion:
          type: string
...
    PublicClaim:
      type: object
      required: [orderId, spotIndex, brandName, amountCents, reservedAt, status]
      properties:
        orderId:
          type: string
        spotIndex:
          type: integer
        brandName:
          type: string
        destinationUrl:
          type: ["string", "null"]
        logoObjectPath:
          type: ["string", "null"]
        amountCents:
          type: integer
        reservedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [reserved, funding, payment_failed, funded]
...